backend/requirements.txt
# === Web Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-multipart==0.0.18

# === Database ===
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
asyncpg==0.30.0

# === Cache & Queue ===
redis==5.2.1
celery==5.4.0

# === Auth ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# === Storage ===
minio==7.2.12

# === Search ===
elasticsearch==8.17.0
pymilvus==2.5.3

# === AI - Vision & NLP ===
paddleocr==2.9.1
paddlepaddle==2.6.2
cn-clip==1.5.1
insightface==0.7.3
onnxruntime==1.20.1

# === AI - Speech ===
funasr==1.2.2

# === AI - Utils ===
pyscenedetect[opencv]==0.6.5
imagededup==0.3.2
pyiqa==0.1.12

# === Image & Video Processing ===
Pillow==11.1.0
opencv-python-headless==4.10.0.84
ffmpeg-python==0.2.0

# === Utils ===
pydantic==2.10.4
pydantic-settings==2.7.1
loguru==0.7.3
httpx==0.28.1
numpy==1.26.4
apscheduler==3.10.4
